safari-534.53-branch/LayoutTests/fast/viewport/viewport-warnings-2-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument key "wwidth" not recognized and ignored.
viewport size 980x1078 scale 0.326531 with limits [0.326531, 5.000000] and userScalable -1.000000
